refactoring and fixed test
