+++ src/main/java/com/mageddo/dnsproxyserver/di/Context.java
@@ -9,6 +9,7 @@ import com.mageddo.dnsproxyserver.di.module.ModuleMain;
 import com.mageddo.dnsproxyserver.di.module.ModuleMap;
 import com.mageddo.dnsproxyserver.di.module.ModuleSolver;
 import com.mageddo.dnsproxyserver.di.module.ModuleStartup;
+import com.mageddo.dnsproxyserver.di.module.SolverRemoteModule;
 import com.mageddo.dnsproxyserver.docker.dataprovider.DockerNetworkFacade;
 import com.mageddo.dnsproxyserver.quarkus.QuarkusConfig;
 import com.mageddo.dnsproxyserver.server.Starter;
@@ -36,7 +37,8 @@ import java.util.Set;
   ModuleSolver.class,
   ModuleStartup.class,
   ModuleMap.class,
-  ModuleConfigDAO.class
+  ModuleConfigDAO.class,
+  SolverRemoteModule.class
 })
 public interface Context {
