+++ src/main/java/com/mageddo/dnsproxyserver/di/module/SolverRemoteModule.java
@@ -1,0 +1,15 @@
+package com.mageddo.dnsproxyserver.di.module;
+
+import com.mageddo.dnsproxyserver.solver.remote.CircuitBreakerService;
+import com.mageddo.dnsproxyserver.solver.remote.application.CircuitBreakerFailSafeService;
+import dagger.Binds;
+import dagger.Module;
+
+import javax.inject.Singleton;
+
+@Module
+public interface SolverRemoteModule {
+  @Binds
+  @Singleton
+  CircuitBreakerService circuitBreakerService(CircuitBreakerFailSafeService impl);
+}
