+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/CircuitBreakerService.java
@@ -1,0 +1,9 @@
+package com.mageddo.dnsproxyserver.solver.remote;
+
+import java.util.function.Supplier;
+
+public interface CircuitBreakerService {
+
+  Result handle(Request req, Supplier<Result> sup);
+
+}
